The very first draft for #1
Developed during period 2023-02-15 - 2023-03-10.

Originally the idea in the issue was to create a mobile app with a name
like "who-is-he-how-is-she-meter".

When I started to work on the implementation I realized that it actually
would do the very basic and useful things to care about the most important
person in user's life. Thus I decided that the name of the app would be
something like "i-care.by <platform> app".

I chose Flutter to write code once and then compile it to different
target platforms.

"i-care.by web app" is available at
https://mshurkaeu-public.github.io/i-care.by-app/web/

In this first commit the following is implemented:
1. User preferences (name and preferred pronoun).
2. Display the list of the diary records on home page.
3. Editing today record (- want to do today, - emotions in that moment,
- done today, - emotions in that moment)
4. Options to answer "who is the most important person in your life" are
randomized on each run, so the user needs to focus and find the
appropriate one.

Each day the application asks (trigerred by user)
1. Who is the most important person in your life?
2. What do you want to do for the person today?
3. What are your emotions and feelings in this moment?
--
then, presumably in the second part of the day,
4. What did you do for the person today?
5. What are you emotions and feelings in this moment?

The answers are recorded and stored in json format.
On web platform - in window.localStorage.
On other platforms - in application documents folder.
----

I want to say thank you to the following people, who helped me to create
this first draft (chronological order, most recent accepted help go first,
sorted by "accepted on" because there was a time lag between a moment the
help was provided and when results appeared in the application (and some of
the already provided hints are not yet incorporated into the application)):

2023-03-10 02 🦸 — provided valuable feedback about my texts.
And that helped to make the texts more clear.

2023-03-10 01 🦸‍♀️ — aspired me to highlight in the application UI that the
"want to do" list is a list of desires and not a list of "duties".

2023-03-06 02 🦸 — aspired me to implement support of the preferred pronoun
setting and thus to implement more user-friendly texts.

2023-03-06 01 🦸‍♀️ — aspired me to implement the ability to edit user's name
via UI sooner.

2023-03-05 02 🦸 — found errors in text on the second introductory screen and
shared a feeling that was not clear what to do after the "want to do today"
list was populated.

2023-03-05 01 🦸‍♀️ — aspired me to think how to implement proper validation for
a list of persons to actually be a list of several persons.

2023-03-04 🦸‍♀️ — shared own vision about the target audience and that helped
me make my own vision of that audience clearer and to reflect that in texts.

2023-02-28 🦸 — found a typo on the second introductory screen and also
aspired me to think about UI element positioning and to make button "Next"
closer to "major events".

2023-02-15 🦸 — recommended me to use Flutter to achieve my goal
"one code -> many target platforms". I hadn't known about Flutter before that.