Replace usage of helperText with hintText in emotions TextField
in order to reduce required space on a device screen.
On my smartphone the helper text eats significant space.

Disadvantage of hintText though is that it dissapears when user inputs
first character into the field. Thus in future, perhaps, I'll introduce
some sort of hint, which will appear when user needs/requests it.