Make onDoneButtonPressed a named parameter of _WhatIsDoneRequestContainer constructor
I want to change parameters in the onDoneButtonPressed function signature.
This will cause cascading modifications in classes, inherited from
_WhatIsDoneRequestContainer. Making the onDoneButtonPressed a named
parameter allows to use dart syntax sugar to re-use definition of a
parameter from the parent class. This will allow to make the modifications
of onDoneButtonPressed function with minimal changes in source code.
Only start and end points in a chain of calls will be affected.