Explicitly specify initial value for autoSubmitFirstScreen local variable in _EditDiaryRecordState.initState
this is a step towards fixing regression, introduced yesterday during
code refactoring. When a new DiaryRecord is created then who field of
that is null initially. As a result of that - code inside
if (_theMostImportantPerson != null) doesn't run. And as a result
line _stagingDiaryRecord = DiaryRecord({}) is not executed. And that
leads to an error like this:
LateError (LateInitializationError: Field '_stagingDiaryRecord@39311160' has not been initialized.)

Thus prepare the code to remove the if (_theMostImportantPerson != null)
So initialization of _stagingDiaryRecord would happen in any case.