Remove "if (_theMostImportantPerson != null)" from _EditDiaryRecordState.initState
this fixes regression, introduced yesterday during code refactoring.
When a new DiaryRecord was created then who field of that was null
initially. As a result of that - code inside
if (_theMostImportantPerson != null) didn't run. And as a result
line _stagingDiaryRecord = DiaryRecord({}) was not executed. And that
leaded to an error like this:
LateError (LateInitializationError: Field '_stagingDiaryRecord@39311160' has not been initialized.)

Now initialization of _stagingDiaryRecord happens in any case.