Use Noto Color Emoji font to display the thinking face
to make the thinking face look the same way on different platforms.
Without this font usage the face looked differently on different
platforms. For example, looked as colored thinking face on Linux
desktop and looked as a grey outline the thinking face on web.

I use the Noto Color Emoji font from
https://github.com/googlefonts/noto-emoji