Workaround a problem with displaying text of "my gratitude" section in Windows browsers
usage of Roboto font family instead of Segoe UI allows to workaround the problem.
I don't understand why there is an issue with the Segoe UI. Is it because Noto Color
Emoji is some different kind of a font?

But the text with the problem on Windows looks so ugly that is unbearable. Thus I
first implement a quick workaround and then find an explanation to the behavior.