Introduce a local variable lightTheme inside method ICareByApp.build
to prepare code to set font family on the application theme level.
This is required because texts, containing mixture of glyphs from
Segoe UI and Noto Color Emoji font families rendered in the very
ugly way. The same happens to mixture of San Francisco and Noto
Color Emojo glyphs.

So the idea is to set Roboto as a major font family and Noto Color
Emoji as a fallback font family. Preventing the default behavior
when major font family depends on a run-time platform. The solution
may be revised after the root cause of the ugly behavior is found.