Do not include Noto Color Emoji font into the application assets
It makes sense to include Noto Color Emoji font into assets only if
a target device can get any benefit from this. For example, if the
font is not available on a target device and it's not reasonable to
load the font via the Internet (via google_fonts package). But if we
target web platform then the font will be loaded to the device via
the Internet in any case. Either from our own web server if we include
this font into the assets. Or from Google web servers if we don't
include the font into assets and use GoogleFonts.notoColorEmoji instead.