Introduce local constant sideButtonPadding inside method _MyPhotoGalleryState.build
the value 8 of the constant matches the default TextButton padding.
However introducation of the explicit value here in the code will
allow perform button width calculation on the further steps.

I'm going to use SizedBox to wrap hint text. To always allow "swipe
buttons" be visible on the screen. Text of the hint is expected to
wrap to multiple lines in some cases.