Rename class _MyPhotoGallery to MyPhotoGallery
to prepare that to move into a separate dart file. To re-use the
MyPhotoGallery class later in other places. I'm going to introduce
a separate page with a description how to find out who is the most
important person in your life. That page will contain text with
inserted images.