Experiment with a zero width space
I suspect there is a bug in Flutter, when widget is inserted into
Text.rich and after that there are two new line characters.
On Linux and in desktop browser that works as expected. But on
mobile browser the empty line between the widget and following text
is absent. Thus I try this experiment to check if a zero wide space
helps workaround the bug.