Enable use_key_in_widget_constructors code analysis rule
because there are no currently places in the project code, where
the rule is violated.

I don't fully understand the nuances behind the rule, but from what
I've read 5 minutes in Flutter source code it seems to be reasonable
to follow the rule. So, will follow the rule until it is proven to
make some harm.