Introduce a link on the second introductory screen to return to the first screen
this is required in preparation of introduction of the drawler on
the second introductory screen. When there is a menu then AppBar
of a Scaffold doesn't show "back" button. As a result as soon as I
introduce menu on the second introductory screen then on some
platforms there will be no way to return to the previous screen and
correct user's name if required.

This is why I need to introduce a way to have an ability to correct
the name by returning to the first introductory screen when required.