Remove date of a thought from the second introductory screen
when I started to write down the thought I was not sure how solid
that would be. Thus decided to set a marker, when I thought that.
Now, I think, the marker just complicates reading of the text to
a user. And actually doesn't provide much value. It is possible
to check git repository to know the date or check version and
build of the application to get some idea if that thought is
relatively new or relatively old.