Description of the very first attempt to build Current Reality Tree
for a dependency between a cause "the great majority of belarusian
citizens do not treat themselves as the most important people in their
own lives" and an effect "dictatorship in Belarus".

Also record my gratitude to Eliyahu M. Goldratt for his book "It's
Not Luck" and my gratitude to the person, who listened to my
description of the first version of the mentioned current reality tree
and provided feedback.