The 223 version of the current reality tree
"Mirror reflect" the CRT. After the experiment with a friend
"CRT as a map of reality" on 2023-11-18 I got some insights and
relaized that we worked with the CRT only as with a "map of reactions
to officials' and special forces' actions". While there was another
part "person other wishes". And I thought that this other part even
more important. That was an intuition. Not a very well developed
decision. Anyway, I followed the intuition and "mirror reflected"
the CRT. So, "wishes" are now on the "1st" (left) side of the schema.

Also record my gratitude to my fried, who agreed to participate in
experiment "CRT as a map of reality" and my gratitude to Stephen R.
Covey for his book "The 7 Habits of Highly Effective People", from
where I got an idea about "maps of reality".