Change approach to labeling versions of CRT
Instead of date and version number on some selected versions put
current date on the first page of CRT document.

This way, when the document will be opened via a link through draw.io
web site it will be clear from the first page when was the last
modification of the document.

This approach will require to not forget to modify the date on the
first page of the document each day, when the document is edited.