Modify wording a bit on the second introductory screen
to reference "the most important person" instead of "that person".
The latter in the text means "the person, who is declared to be the
most important person". Because I suspect, that in reality many people
sincerely think that someone is the most important person in his/her
life. But actually they care about themselves and not about the
person, who they declare to be "the most important person in their
life".

This modification, perhaps makes no difference for a reader, but
for me there is a significant difference. I tell what I really
think, and do not state that some people are liars in the text.
Because I don't know what is actually in their minds and pictures
of the world.