Update text in the pink area on CRT to match title of the hypothesis 2
https://github.com/mshurkaeu-public/i-care.by/discussions/5

I'm going to use a separate layer with a pink area for each hypotheses.
The pink area will highlight part of CRT, which is related to a
corresponding hypothesis. And to be sure what the corresponding
hypothesis is, while navigating the CRT, it makes sense to have
the hypothesis title in the pink area.

Also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 23.1.1 of https://draw.io.