Introduce a new grouping circle on a line between
"in existing circumstances officials and special forces perform some
specific action towards society" node and "all possible types of
citizen's attitude to the specific action of officials and special
forces towards society". I need that circle because I want to make
more explicit on the diagram that "ЧР.1" and "ЧР.2" are also subject
of similar attitude. I'll connect the two mentioned nodes and the
just created circle.