Move first large block of text on the introductory page for the CRT
to be as high in the hierarchy of xml document elements as possible.
I need this to move "date" element as high in the xml document
elements hierarchy as possible. And that I need to have "modified"
attribute of mxfile element and value stored in the "date" element
as close as possible one to another. So, when "modified" attribute
has a new date then higher chances to notice that value stored in
"date" element is out of sync.