Describe some possible causes for "a citizen doesn't recognize an obstacle, which stops him/her" effect
I think I'm getting close to a moment when I'll be able to connect
hypothesis 1 with its "cause" hypothesis 2.

Intuitively: A citizen has no reason to figure out/learn in which
parts of reality he/she exist. Or by which parts of the reality
he/she is affected. If he/she is not the most important person in
his/her life then what's the point to figure out all the aspects
of the own life? Instead need to pay "all attention" to the most
important person.

Today I think:

It's like driving a car (rather piloting space ship), but looking
at the state of driver/pilot in a neighbor car/space ship instead
of figuring out what aspects of driving the car/space ship exist at
all. And what those aspects mean/affect.

State of a neighbor car/space ship and its driver/pilot of course
matters. But this cannot be single priority in the driving/piloting.
And declaring that "neighbor driver/pilot is the most important person
in your life" is likely leads to ignoring completely critical areas
in steering own car/space ship.

I think we currently leave in the "crazy carousel" world, where
most of the drivers/pilots do not look in the window/at the radar
and instead look what's with the neighbor driver/pilot. Not ensuring
it's safe to do this in current road/space situation.

We'll see, how deep the rabbit hole goes...