Modify statement of the 2nd basic hypothesis
I realized a need for the "a citizen is conscious" or another
statement like that to discuss if the citizen knows that something
is required to do. If one is unconscious then it is difiicult to
judge if the person knows something or not.

That realization happened to me approximately 10 days ago, while I
was working on the causes why a citizen cannot do something even
though the citizen knows it is required to do/may be done and knows
how to do that.

Also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of https://draw.io version 23.1.5. I use that web
application to draw CRT for dictatorship in Belarus.