Replace word "run away" with word "escape" on the diagram
I believe the "escape" is more wide term and thus better suits the
diagram.

Yesterday I saw "escape" in some Google search results when tried
to search for a list of possible reacions on something new/unknown.
Didn't find what I wanted to find, but noticed word "escape" and
today decided to apply that on the diagram.