Introduce a new possible "effect" of "internal process of decision making" node
"The initial wish is not attractive anymore". During the iterative
process it is possible that the citizen learns something which
makes the initial wish not attractive anymore.

Just a stupid example to illustrate:
a person gets to an inhabited island and wants to eat a burger.
Decides to cook the burger "from scratch" for the first time in real
life and during the process of struggling with obstacles (described
on the diagram) figures out that he/she needs to catch and kill an
animal for that. And the person learns about himself/herself that
he/she doesn't want to kill animals (somehow has never thought
about that requirement before the moment). So the original wish may
become not attractive anymore. And the person decides to become a
vegetarian since that moment. To be not related to killing animals.