Introduce and use placeholder "modifiedDate" in the title "stamp"
there is no standard placeholder for the modified date. It's a pity
because there is "modified" attribute on mxfile node in drawio file.
Would be much more convinient to use that value. But… we have what
we have.

Another problem is that proerty can be defined only on page level.
And global variables are defined on application level and saved in
browser storage between application run.

Thus I'm going to introduce the property on each separate page and
update value of the property via script. To have synced "stamps" on
all pages.