Update comment for "internal process of decision making" node
I added "Ignore" to the list of possible outcomes. And introduced
"System failure" outcome. And created sublist of "System failure"
options: "Paralysis", "Loss of consciousness", "Death".

Also fixed a coupl of typos in the text of the comment.

Also record my dratitude to the person, who helped me realize that
"Ignore" is absent.