Rename "knowledge base" node to "informational state"
I want this to be similar to "emotional state" and "physical state".
Because if I'll manage to explain why it makes sense to care about
emotional and physical state then it will be easier to understand
why the same will be applicable to informational state. Similarity
of names may help.