Fix an issue/error in basic hypothesis 2 "there are only 4 reasons why a person doesn't do something"
that was "there are only 5 reasons why a person doesn't do something".
But during my recent work on the description of "learned helplessness"
I realized that a reason "doesn't know how to do something" is just
an obstacle on a way of the person. Thus I realized that it should
be removed from a list of reasons in the basic hypothesis 2.

And also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 24.4.13 of website https://draw.io.