Record my gratitude to Alexander Orlov and school of management "Stratoplan"
for a hint about "4 reasons why a person doesn't do something". On
my today opinion that was quite wrong opinion, but that was a good
start for my first attempts to describe reasons why a citizen does
nothing in response to authority impact.

I think that is related to commit cf240edf875c4ef3a25de8a397f5fab91aea4fff.
And the link to look at the CRT diagram in that moment is
https://viewer.diagrams.net/index.html?edit=_blank#Uhttps://raw.githubusercontent.com/mshurkaeu-public/i-care.by/cf240edf875c4ef3a25de8a397f5fab91aea4fff/current-reality-tree/Дрэва%20цяперашняй%20рэальнасці.drawio