Modify all columns with reasons based on basic hypothesis 2
to set make z-order of the nodes in the list to match visual
order of the nodes. To match order of reasons in the description
of the hypothesis 2.

And also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 24.5.2 of website https://draw.io.