Move "loss of consciousness" node closer to "internal process of decision making" node
this is preparation to move "internal process of decision making"
node and most of the related to it nodes further to the right on
the diagram. In order to reduce number of lines intersections and
to make the diagram more clear.

And also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 24.5.3 of website https://draw.io.