Modify text of possible reasons of "wants to achieve the initial goal and doesn't try to change the reality" node
to use verb "realize" instead of "know". Currently it's not clear
to me if "knowing" something is sufficient in decision making.
If there is something in my memory or in my DNA or in my records
on paper and don't currently use this information, just ignore that
as that was never "known" to me then what's the difference if I
know that or not. Perhaps "realize" is better. I think that
"realize" is something like "is currently in CPU register", information
that is being used actively.

It is kind of different grades if "know":
0. There is something that is in the reality. If that is in reality
then I theoretically can extract that information and use that in
my decisions.

1. There is something that is in my memory. This is something that
I recorded into my memory as truth about reality told me by somebody
or perhaps I extracted that information about the reality myself and
recorded that into my memory. I theoretically can extract that record
from my memory and use that in my decision.

2. There is something in my "register". Information that I use right
now to make my decision. I theoretically can make a mistake in lose
it even from there. But the chances to use that piece of information
are higher. I believe this is what I understand as "I realize that
in the current moment".