Add missing logical "and" operations on all the possible causes of "wants to achieve the goal but doesn't try to change the reality" node
the "ands" had to be added in the commit a6022c46eeb82785f5c1fa291ef640cef34d0c2c
But in that commit I made a mistake and forgot about the "ands" even
though I remember I thought about them in some moment. Since it's
already 3rd after the commit with the mistake I do not want to
edit the commit and want that to demonstrate that I made the mistake.