Move comment for node "one or more of real obstacles are nature laws" from "effect" to "cause" of the "wants to achieve the initial goal and doesn't try to change the reality" node
preserving the text of the comment as is. In order to get prepare
the diagram to get rid of the "effect" copy of the "cause". And
the same text most likely will highlight the logical mistakes.
That comment on the "cause" looks a bit akward. Perhaps this is
an early manifistation of logical error.