Replace "causes" of the "wants to achieve the initial goal and doesn't try to change the reality" with standard "causes" according to the basic hypothesis 2
because that reasonable to describe why a person doesn't do something
using the basic hypothesis 2.

It seems that the "learned helplessness" has to be connected to the
"doesn't want to do" "cause". Because "cannot do" "cause" is
described recursively via ➀ template and "doesn't know" "causes"
seems to be not applicable at all in this context. "Follow a desire"
should present on some very low level of "alive being implementation".