Modify text of the "there is a set of obstacles A₁ in the reality"
to use the exact phrase "which block the citizen from executing
action D" instead of the "which block the citizen from achieving
his goal". The old version of the text implicitly changed context.
And I started to think about some "arbitrary goal" instead of keeping
focus that the "goal" was to "perform action D".

I hope more specific wording will help to figure out what is missing
and what is wrong in this part of the CRT.