Introduce "branching": either "tries to perform action D" or "doesn't try to perform action D" after the "doesn't realize there are obstacles" node
it's not intuitive that if a person wants to perform action D, doesn't
realize there are obstacles for that and still doesn't perform
action D. But to cover all the reality all options should be included.
And after the inclusion I start to think that I met something like
this in my own life. Cannot describe it yet, but I feel like I met
that. It's like when you have contradictory desires and just don't
recognize one desire to be an obstacle for another. Perhaps, such
situation is an indicator of not sufficient awarness.