Amend text in nodes mentioning set of imagined obstacles Bₙ
to match recent update on the description of real obstacles Aₙ.
I think this updated wording kind of relaxes imagined "strictness"
of the obstacles. This is just my perception. Not sure this can
be signtifically measured. On my opinion the new wording makes
imagined "average reader" to easier understand/suppose that the
obstacles can be overcome. 🤷