Move "death" and "loss of consciousness" "effects" out of the description "a person cannot do something" part of the diagram
because those two are first reasosn of the list of reasons why a
person doesn't do something. The moved two can be treated as  the
basic reasons for the rest of the reasons from the same list
(doesn't know it's possible/required to do action D; doesn't want
to perform action D; cannot perform action D).

And also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 24.7.4 of website https://app.diagrams.net.