Amend text of "causes" of "wants to perform action D and doesn't try to change the reality" "effect"
to mention explicitly that this is situation when the citizen
wants to perform action D, realizes there are some obstacles for
that and doesn't try to change the reality. I believe this
clarification will help to analyze and to detect deeper "causes".

And also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 24.7.6 of website https://app.diagrams.net.