Move the highlighted area for hypothesis 1 about the root cause of the dictatorship in Belarus
to highlight part of the "causes" of "has confidence that he cannot
change the reality in the direction of performing action D".

Just a note: as far as I remember, this is the third move of this
area during my work on this diagram. :) But each time, I think, the
highlighted area is more and more proper one.