Upgrade Flutter package carousel_slider to version 5.0.0
in order to resolve the compilation error, which happens if I try
to upgrade Flutter to version 3.24.0. The error and it's reasons
were described in commit f86f9f87a3d7ac89b9f70cc4ebfc226f75ec09f1.

And also record my gratitude to the author of the package serenader.me
for creation of the package version.