Introduce a "split of reality":
"has some opinion about own ability to change reality in the direction
of performing of action D" and
"has no any opinion about own ability to change reality in the
direction of performing of action D".

This helps describe reality fully. Without the split a piece of
reality was missing on the diagram.