Introduce a comment for "formula" why a citizen cannot do something
to explain the formula better to people, who don't understand that
formula from the first glance and start arguing telling that sometimes
a person invent unexisting obstacles (for example to explain to
himself and to others why the person cannot do something). I
believe the new comment will help to explain to such people that
the formula is still valid. They just need to look from a bit
"higher" point or from another "angle".

And also record my gratitude to the people to helped me realize
existance of such people and helped me unesrtand the reason of
those people misunderstanding.