Remove a note that "I don't know how to connect realization of obstacles existance to the perception of reality" from the comment for "perception of reality" node
because I've already connected the nodes somehow. Not perfectly
because arrows on the diagram mean "this causes that" and here
the connection arrows means "this is a part of that". I'm not
quite happy with the approach, but … will use it for now. Perhaps
later I'll find some better way to describe reality on the diagram.