Introduce a "merge point" for all internal and external circumstances of a person
to be able to move the list of circumstances further to the right
and to not affect shape of outgoing paths from the curcumstances
nodes. The list will be move together with the "merge point" to
free up horizontal space to describe "internal process of decision
making" in a more detailed way.