Replace usage of "important information" with "information important for the person" "term"
to better describe what I mean. I want to explore reasons why a
person may ignore information, which would be treated as "important"
by himself/herself. In reality there is a lot of incoming information
and I suspect it's physically impossible (or at least difficult for
currently living people) to recognize and deeply process all of
that. But if it happens that the person himself/herself would
treat some information as important, but ignoring that may be
an interesting case.