Amend statement about state of relationhips between society and (officials and special forces)
to mention that effect of 3rd parties on the both sides of the
relationships also matters.

Also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 24.8.3 of website https://app.diagrams.net/.

And also record my gratitude to the person, who helped me to
understand that it makes sense to mention "3rd party impact" in
the statement of the state of relationships between society and
(officials and special forces).