Introduce a comment for "state of relationship between the society and (officials and special forces) is a logical result of behavior of both side and of the impact on the sides of "3rd forces"
to describe what I mean under "3rd forces".

And record my gratitude to the people who helped me to understand
that it makes sense to do that.

And also record my gratitude to JGraph Ltd and draw.io AG teams
for creation of version 24.8.4 of website https://app.diagrams.net.