Replace description of my thoughts about non-linear nature of impact of number of citizens to officials and special forces with description of possible examples of a single citizen impact
reality showed that there are people who cannot grasp that as
a side-comment during discussion of "statement 2": impact of the
society is somehow combined from impacts of individual citizens.

I think, I'll need to return to those thought much later, when the
root cause of the dictatorship is established much more precisely
than today. That may become a part of the future reality tree
diagram. 🤷