Inroduce a missing node for not described part of reality when a citizen impacts officials and special forces without having some requieremt to them
kind of "side effect" impact: a teacher, who teaches children in
class impacts those children and some of them will become officials
and special forces. Also some of the children are children of
officials and special forces and thus the teacher indirectly may
impact the parents.

Another example - a boy playing with a ball on a street may break
someone's window. And that could be a house of some official or a
person from special forces. Thus that boy made some impact on
officials and special forces.

"Butterfly effect" is possible 🤷.