Introduce a statement 0 "in order to understand each other we need to talk the same language"
I met a couple of situations when hints to the problem appeared.
But one case was especially bright. We could not agree about meaning
of the word "улада". I record my gratitude to the person, who
helped me to realize how important is to talk the same language.

And also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 24.9.1 of website https://app.diagrams.net/.