Replace "informs officials and special forces about his wish" "effect" with "kills his requierement to officials and special forces" "effect"
because I realized that this part of description of reality is not
symmetrical with the description of reaction to impact of officials
and special forces. And thus I understood that something is missing.

Perhaps this is not a final version. It is interesting that it is
easier for me to describe reactions to external impact than to
describe reactions on internal impact. Currently I do this rather
looking for the symmetry than looking "inside".