Move a comment with my thoughts about dictatorship in Belarus 280 points down
to free up vertical space to insert a new comment with thoughts
about belarusian word "уздзеянне" (impact). I need that because I
found people, who don't recognize impact with strength zero as a
special case of impact. That case is kind of "completely different"
thing in their private picture of the world. And that complicates
understanding of generic approaches for them.

And also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 25.0.1 of website https://app.diagrams.net.