Decrease height of a comment for the "path" to the list of all possible options how a citizen may perceive his own requierement to official and special forces by 40 points
to free up vertical space for an experiment. I want to try to move
down lists of types of impact of a citizen to officials and special
forces for cases when a requierement appeared by reasons other
than impact of offcials and special forces to the society.

I want to try to split a list of all possible cases when a citizen
impacts officials and special forces into two lists. That is
expected to simplify discussions of the diagram with other people.

And also record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 25.0.3 of website https://app.diagrams.net.