Introduce short description of 1st level "causes" of "a person knows that it's possible/required to try to change reality in the direction of performing of action D but doen't want to try" node
I kind of wanted to express that thought long time ago, but required
words didn't come. 🤷 A conversation with a person about imaginary
person SP₂ helped me to finally realize what I want to express and
to find the words.

Record my gratitude to the person, who discussed the topic with me.