Amend text of 1st level of "causes" why a person doesn't want to try to change reality in the direction of performing of action D
because "to know" an option doesn't mean "to consider" the option.
For example in school I kind of "knew" that magnet worked
differently for steel and aluminum. But didn't consider the option
to distinguish painted balls of equal size and weight from steel
and aluminum.

My current understanding is that "knowledge" is some information
available in your own neural system. But during the process of
decision making that information could be just not used. As the
result we have "I knew but didn't consider/use that" situaton.