Amend text of "doesn't consider (and doesn't try) other ways to try to change reality to approach to performing of action D" node
to use present tense instead of the past tense and to use "to approach"
instead of "in direction". Because I switched today to use word
"direction" to mention the 5 possible *directions* to try to change
reality *to approach* performing of action D.

And also introduce a comment to the modified node to explain why
this particular statement is required in this place of the diagram.