Replace phrase "in the direction of performing of action D" with phrase "to approach performing of action D" on multiple nodes and comments
to have more sutable description of reality. Sometimes approaching
performing of action D may be via "part of path in ortogonal direction".
For example "workaround some obstacle" may be more effective then
"straight through the obstacle" (but in the *direction* of performing
of action D).

And also in this commit there is a fix of a typo and replacement
of "the 5 directions to try to change reality" phrase with "the 5
categories of ways to try to change reality". I decided that usage
of word "direction" is not suitable enough here. Those are neither
"directions" nor "dimensions" but some kind of categories. Maybe
better name will come later. 🤷