Modify texts of possible elements of behavior in reaction to the partciular impact of officials and special forces
to describe the elements in terms, which are valid for any entity
met by a person. I realized today that these three elements are
generic and valid for any situation when a person meets some
entity. After that the "all pieces" of the person behavior will
fit into the set: "increases chances of the entity to exist",
"decreases chances of the entity to exist", "doesn't affect chances
of the entity to exist". And the entity may be absolutely everything.
Nature law, any object, any person, any animal, any phenomenon.