Amend text of "a requirement to officials and special forces appears and the citizen imagines that requirement supports implementation of the imagined better future" node
I'm trying to find appropriate wording, which will explain what's
going on in reality. This modification explicitly states, that
requirement appears as a desired path to move to the imagined
better future.

And record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 26.0.4 of website https://app.diagrams.net.