Amend text of nodes describing every citizen and every person from officials and special forces
I'm trying to understand if logical "and" is required on the incoming
paths. With the new texts it becomes more clear to me, that the
logical "and" make sense. Because now the nodes state the forces
"create some circumstances in current moment". Which now sounds
more reasonable than the statement "result of the 3rd party forces
and relationships between the society and `officials and special
forces` is the person lives own life how the person can".