Amend comment for "third forces" nodes on the page 3 of the document
this updated text helps me realize what it missing of the diagram.
I'll introduce "aspects of relationships with the society and
`officials and special forces` which create circumstances for the
citizen" nodes. Or something like that. To better highlight that
the citizen makes choices in the circumstances, created by the
"third forces" and by other people.

And that in it's turn will highlight a need to introduce
"circumstances of life created by the person himself/herself".