Describe essence of having a requirement to `officials and special forces` via execution of will
since dictatorship is a situation when one party subordinates another
party to it's will then to find a root cause of the effect I need to
find a way to describe processes on both sides via execution of will.
Sounds simple, but somehow doesn't happen fast and easy. I guess,
I have not enough intuition to do this fast. Thus I do it slowly :)
and build better understanding of reality during the provess.

And record my gratitude to JGraph Ltd and draw.io AG teams for
creation of version 26.0.6 of website https://app.diagrams.net.