Introduce "statement 3"
I've heard from one person, that statement "the state of relationships
between society and `officials and special forces` is natural outcome
of behaviour of the sides and impact on them from 3rd forces" may be
not intuitive for an "average person". Well… I haven't checked if
that is true or not. Rather true… if there is at least one such
"average person" then the statement is true. So, I hope the
statement 3 will be a doable step for such a person.